argc: refactor: move semamtic analyser to seperate file
